refactor show page styles
