+++ brewery-app/src/styles/ShowPageStyles.js
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+
+export const ShowPageContainer = styled.div`
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: column;
+    align-items: center;
+`
+
+export const BrewDiv = styled.div`
+
+`
+
+export const BrewImage = styled.img`
+
+`
+
+export const BrewStats = styled.div`
+
+`
+
+
+export const BreweryName = styled.div`
+    
+`
